ingore

+++ app/src/main/java/com/habbyge/iwatch/util/FileUtil.java
@@ -1,15 +1,10 @@
 package com.habbyge.iwatch.util;
 
-import android.util.Log;
-
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.nio.channels.FileChannel;
-import java.util.Enumeration;
-import java.util.jar.JarEntry;
-import java.util.jar.JarFile;
 
 /**
  * Created by habbyge on 2021/1/5.
@@ -42,20 +37,13 @@ public class FileUtil {
 		}
 		if (file.isDirectory()) {
 			File[] files = file.listFiles();
+			if (files == null) {
+				return true;
+			}
 			for (File f : files) {
 				deleteFile(f);
 			}
 		}
 		return file.delete();
-	}
-
-	public static Enumeration<JarEntry> parseJarFile(File jarInputFile) {
-		try {
-			JarFile jarFile = new JarFile(jarInputFile); // JarFile
-			return jarFile.entries();
-		} catch (Exception e) {
-			Log.e("iWatch.FileUtil", "parseJarFile, exception: " + e.getMessage());
-			return null;
-		}
 	}
 }
